Move the FormatterTestHelper to correct project and make all test cases in CollectionTest.cs pass
